Begin refactor of the code with coding partner William, and reorganize the structure of the files. Everything related to Books is already refactored.
