+++ classes/person.rb
